--- a/docs/en/guides/queue-management.md
+++ b/docs/en/guides/queue-management.md
@@ -39,40 +39,35 @@ Queue management using [Ultralytics YOLO11](https://github.com/ultralytics/ultra
 
         ```python
         import cv2
-
-        from ultralytics import YOLO, solutions
-
-        model = YOLO("yolo11n.pt")
-        cap = cv2.VideoCapture("path/to/video/file.mp4")
-
+        from ultralytics import solutions
+        
+        cap = cv2.VideoCapture("Path/to/video/file.mp4")
+        
         assert cap.isOpened(), "Error reading video file"
         w, h, fps = (int(cap.get(x)) for x in (cv2.CAP_PROP_FRAME_WIDTH, cv2.CAP_PROP_FRAME_HEIGHT, cv2.CAP_PROP_FPS))
-
+        
         video_writer = cv2.VideoWriter("queue_management.avi", cv2.VideoWriter_fourcc(*"mp4v"), fps, (w, h))
-
+        
         queue_region = [(20, 400), (1080, 404), (1080, 360), (20, 360)]
-
+        
         queue = solutions.QueueManager(
-            names=model.names,
-            reg_pts=queue_region,
-            line_thickness=3,
+            model="yolo11n.pt",
+            region=queue_region,
         )
-
+        
         while cap.isOpened():
             success, im0 = cap.read()
-
+        
             if success:
-                tracks = model.track(im0, persist=True)
-                out = queue.process_queue(im0, tracks)
-
+                out = queue.process_queue(im0)
                 video_writer.write(im0)
                 if cv2.waitKey(1) & 0xFF == ord("q"):
                     break
                 continue
-
+        
             print("Video frame is empty or video processing has been successfully completed.")
             break
-
+        
         cap.release()
         cv2.destroyAllWindows()
         ```
@@ -81,53 +76,47 @@ Queue management using [Ultralytics YOLO11](https://github.com/ultralytics/ultra
 
         ```python
         import cv2
-
-        from ultralytics import YOLO, solutions
-
-        model = YOLO("yolo11n.pt")
-        cap = cv2.VideoCapture("path/to/video/file.mp4")
-
+        from ultralytics import solutions
+        
+        cap = cv2.VideoCapture("Path/to/video/file.mp4")
+        
         assert cap.isOpened(), "Error reading video file"
         w, h, fps = (int(cap.get(x)) for x in (cv2.CAP_PROP_FRAME_WIDTH, cv2.CAP_PROP_FRAME_HEIGHT, cv2.CAP_PROP_FPS))
-
+        
         video_writer = cv2.VideoWriter("queue_management.avi", cv2.VideoWriter_fourcc(*"mp4v"), fps, (w, h))
-
+        
         queue_region = [(20, 400), (1080, 404), (1080, 360), (20, 360)]
-
+        
         queue = solutions.QueueManager(
-            names=model.names,
-            reg_pts=queue_region,
-            line_thickness=3,
+            model="yolo11n.pt",
+            classes=3,
         )
-
+        
         while cap.isOpened():
             success, im0 = cap.read()
-
+        
             if success:
-                tracks = model.track(im0, persist=True, classes=0)  # Only person class
-                out = queue.process_queue(im0, tracks)
-
+                out = queue.process_queue(im0)
                 video_writer.write(im0)
                 if cv2.waitKey(1) & 0xFF == ord("q"):
                     break
                 continue
-
+        
             print("Video frame is empty or video processing has been successfully completed.")
             break
-
+        
         cap.release()
         cv2.destroyAllWindows()
         ```
 
 ### Arguments `QueueManager`
 
-| Name             | Type             | Default                    | Description                                                                      |
-| ---------------- | ---------------- | -------------------------- | -------------------------------------------------------------------------------- |
-| `names`          | `dict`           | `model.names`              | A dictionary mapping class IDs to class names.                                   |
-| `reg_pts`        | `list of tuples` | `[(20, 400), (1260, 400)]` | Points defining the counting region polygon. Defaults to a predefined rectangle. |
-| `line_thickness` | `int`            | `2`                        | Thickness of the annotation lines.                                               |
-| `view_img`       | `bool`           | `False`                    | Whether to display the image frames.                                             |
-| `draw_tracks`    | `bool`           | `False`                    | Whether to draw tracks of the objects.                                           |
+| Name         | Type   | Default                    | Description                                          |
+|--------------|--------|----------------------------|------------------------------------------------------|
+| `model`      | `str`  | `None`                     | Path to Ultralytics YOLO Model File                  |
+| `region`     | `list` | `[(20, 400), (1260, 400)]` | List of points defining the queue region.            |
+| `line_width` | `int`  | `2`                        | Line thickness for bounding boxes.                   |
+| `show`       | `bool` | `False`                    | Flag to control whether to display the video stream. |
 
 ### Arguments `model.track`
 
@@ -151,21 +140,19 @@ import cv2
 
 from ultralytics import YOLO, solutions
 
-model = YOLO("yolo11n.pt")
 cap = cv2.VideoCapture("path/to/video.mp4")
 queue_region = [(20, 400), (1080, 404), (1080, 360), (20, 360)]
 
 queue = solutions.QueueManager(
-    names=model.names,
-    reg_pts=queue_region,
-    line_thickness=3,
+    model="yolo11n.pt",
+    region=queue_region,
+    line_width=3,
 )
 
 while cap.isOpened():
     success, im0 = cap.read()
     if success:
-        tracks = model.track(im0, show=False, persist=True, verbose=False)
-        out = queue.process_queue(im0, tracks)
+        out = queue.process_queue(im0)
         cv2.imshow("Queue Management", im0)
         if cv2.waitKey(1) & 0xFF == ord("q"):
             break
@@ -207,9 +194,9 @@ Example for airports:
 ```python
 queue_region_airport = [(50, 600), (1200, 600), (1200, 550), (50, 550)]
 queue_airport = solutions.QueueManager(
-    names=model.names,
-    reg_pts=queue_region_airport,
-    line_thickness=3,
+    model="yolo11n.pt",
+    region=queue_region_airport,
+    line_width=3,
 )
 ```