--- a/docs/en/guides/queue-management.md
+++ b/docs/en/guides/queue-management.md
@@ -39,35 +39,36 @@ Queue management using [Ultralytics YOLO11](https://github.com/ultralytics/ultra
 
         ```python
         import cv2
+
         from ultralytics import solutions
-        
+
         cap = cv2.VideoCapture("Path/to/video/file.mp4")
-        
+
         assert cap.isOpened(), "Error reading video file"
         w, h, fps = (int(cap.get(x)) for x in (cv2.CAP_PROP_FRAME_WIDTH, cv2.CAP_PROP_FRAME_HEIGHT, cv2.CAP_PROP_FPS))
-        
+
         video_writer = cv2.VideoWriter("queue_management.avi", cv2.VideoWriter_fourcc(*"mp4v"), fps, (w, h))
-        
+
         queue_region = [(20, 400), (1080, 404), (1080, 360), (20, 360)]
-        
+
         queue = solutions.QueueManager(
             model="yolo11n.pt",
             region=queue_region,
         )
-        
+
         while cap.isOpened():
             success, im0 = cap.read()
-        
+
             if success:
                 out = queue.process_queue(im0)
                 video_writer.write(im0)
                 if cv2.waitKey(1) & 0xFF == ord("q"):
                     break
                 continue
-        
+
             print("Video frame is empty or video processing has been successfully completed.")
             break
-        
+
         cap.release()
         cv2.destroyAllWindows()
         ```
@@ -76,35 +77,36 @@ Queue management using [Ultralytics YOLO11](https://github.com/ultralytics/ultra
 
         ```python
         import cv2
+
         from ultralytics import solutions
-        
+
         cap = cv2.VideoCapture("Path/to/video/file.mp4")
-        
+
         assert cap.isOpened(), "Error reading video file"
         w, h, fps = (int(cap.get(x)) for x in (cv2.CAP_PROP_FRAME_WIDTH, cv2.CAP_PROP_FRAME_HEIGHT, cv2.CAP_PROP_FPS))
-        
+
         video_writer = cv2.VideoWriter("queue_management.avi", cv2.VideoWriter_fourcc(*"mp4v"), fps, (w, h))
-        
+
         queue_region = [(20, 400), (1080, 404), (1080, 360), (20, 360)]
-        
+
         queue = solutions.QueueManager(
             model="yolo11n.pt",
             classes=3,
         )
-        
+
         while cap.isOpened():
             success, im0 = cap.read()
-        
+
             if success:
                 out = queue.process_queue(im0)
                 video_writer.write(im0)
                 if cv2.waitKey(1) & 0xFF == ord("q"):
                     break
                 continue
-        
+
             print("Video frame is empty or video processing has been successfully completed.")
             break
-        
+
         cap.release()
         cv2.destroyAllWindows()
         ```
@@ -112,7 +114,7 @@ Queue management using [Ultralytics YOLO11](https://github.com/ultralytics/ultra
 ### Arguments `QueueManager`
 
 | Name         | Type   | Default                    | Description                                          |
-|--------------|--------|----------------------------|------------------------------------------------------|
+| ------------ | ------ | -------------------------- | ---------------------------------------------------- |
 | `model`      | `str`  | `None`                     | Path to Ultralytics YOLO Model File                  |
 | `region`     | `list` | `[(20, 400), (1260, 400)]` | List of points defining the queue region.            |
 | `line_width` | `int`  | `2`                        | Line thickness for bounding boxes.                   |
@@ -138,7 +140,7 @@ Here's a minimal example:
 ```python
 import cv2
 
-from ultralytics import YOLO, solutions
+from ultralytics import solutions
 
 cap = cv2.VideoCapture("path/to/video.mp4")
 queue_region = [(20, 400), (1080, 404), (1080, 360), (20, 360)]