--- a/ultralytics/solutions/solutions.py
+++ b/ultralytics/solutions/solutions.py
@@ -78,7 +78,9 @@ class BaseSolution:
         """Initialize the counting region and line segment based on config."""
         self.region = [(20, 400), (1080, 404), (1080, 360), (20, 360)] if self.region is None else self.region
         self.r_s = Polygon(self.region) if len(self.region) >= 3 else LineString(self.region)  # region segment
-        self.l_s = LineString([(self.region[0][0], self.region[0][1]), (self.region[1][0], self.region[1][1])])  # line segment
+        self.l_s = LineString(
+            [(self.region[0][0], self.region[0][1]), (self.region[1][0], self.region[1][1])]
+        )  # line segment
 
     def display_output(self, im0):
         """