@glenn-jocher This PR will only update the `queue-management` solution. All other solutions will continue to function as they currently do. I'll open separate PRs for left solutions that need to be optimized :)

### Main changes
- The Model class will be instantiated within the solutions, so users only need to provide the model file path.
- Object tracking will now be handled within the solutions, instead of being managed by an external video processing loop.

### New Code
```python
import cv2
from ultralytics import solutions

cap = cv2.VideoCapture("Path/to/video/file.mp4")

assert cap.isOpened(), "Error reading video file"
w, h, fps = (int(cap.get(x)) for x in (cv2.CAP_PROP_FRAME_WIDTH, cv2.CAP_PROP_FRAME_HEIGHT, cv2.CAP_PROP_FPS))

video_writer = cv2.VideoWriter("queue_management.avi", cv2.VideoWriter_fourcc(*"mp4v"), fps, (w, h))

queue_region = [(20, 400), (1080, 404), (1080, 360), (20, 360)]

queue = solutions.QueueManager(
    model="yolo11n.pt",
    region=queue_region,
)

while cap.isOpened():
    success, im0 = cap.read()

    if success:
        out = queue.process_queue(im0)
        video_writer.write(im0)
        if cv2.waitKey(1) & 0xFF == ord("q"):
            break
        continue

    print("Video frame is empty or video processing has been successfully completed.")
    break

cap.release()
cv2.destroyAllWindows()
```

### Output Sample

```CLI
Ultralytics Solutions: ✅ {'model': 'yolo11n.pt', 'region': None, 'line_width': 2, 'show': False, 'show_in': True, 'show_out': True, 'classes': None, 'up_angle': 145.0, 'down_angle': 90, 'kpts': [6, 8, 10], 'colormap': None}

0: 384x640 7 persons, 191.1ms
Speed: 10.0ms preprocess, 191.1ms inference, 12.0ms postprocess per image at shape (1, 3, 384, 640)
```

## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Refactor Queue Management in Ultralytics YOLO for improved usability and efficiency.

### 📊 Key Changes
- Refactored `QueueManager` class to inherit from `BaseSolution`.
- Removed redundant imports and streamlined code structure.
- Simplified function calls, removing explicit track extraction in some snippets.
- Updated QueueManager initialization and method signatures to be more flexible.

### 🎯 Purpose & Impact
- **Improved Code Clarity**: Cleaner structure makes it easier for developers to understand and maintain the code. 🛠️
- **Flexibility**: Allows easier customization of queue management tasks by leveraging a base solution class.
- **Performance**: Streamlined processes can potentially enhance real-time video processing performance. 🚀

This update enhances the developer experience by modernizing and simplifying the existing queue management setup.